# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 808 All Purpose Cleaners Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shifts:
    get:
      operationId: listShifts
      tags: [shifts]
      summary: List shifts
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of shifts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shift"
    post:
      operationId: createShift
      tags: [shifts]
      summary: Create a shift
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShiftBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
  /shifts/{id}:
    patch:
      operationId: updateShift
      tags: [shifts]
      summary: Update a shift
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShiftBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shift"
    delete:
      operationId: deleteShift
      tags: [shifts]
      summary: Delete a shift
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /receipts:
    get:
      operationId: listReceipts
      tags: [receipts]
      summary: List receipts
      responses:
        "200":
          description: List of receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Receipt"
    post:
      operationId: createReceipt
      tags: [receipts]
      summary: Create a receipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReceiptBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Receipt"
  /receipts/{id}:
    delete:
      operationId: deleteReceipt
      tags: [receipts]
      summary: Delete a receipt
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
  /todos:
    get:
      operationId: listTodos
      tags: [todos]
      summary: List todos
      responses:
        "200":
          description: List of todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      operationId: createTodo
      tags: [todos]
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
  /todos/{id}:
    patch:
      operationId: updateTodo
      tags: [todos]
      summary: Update a todo
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
    delete:
      operationId: deleteTodo
      tags: [todos]
      summary: Delete a todo
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
...
    Shift:
      type: object
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, employeeId, date, startTime, endTime, createdAt, updatedAt]
    CreateShiftBody:
      type: object
      properties:
        employeeId:
          type: integer
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: string
      required: [employeeId, date, startTime, endTime]
    UpdateShiftBody:
      type: object
      properties:
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        notes:
          type: ["string", "null"]
...
    Receipt:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: ["integer", "null"]
        amount:
          type: string
        paymentMethod:
          type: string
        paymentDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, amount, paymentMethod, paymentDate, createdAt, updatedAt]
    CreateReceiptBody:
      type: object
      properties:
        invoiceId:
          type: integer
        amount:
          type: string
        paymentMethod:
          type: string
        paymentDate:
          type: string
        notes:
          type: string
      required: [amount, paymentMethod, paymentDate]
...
    Todo:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        completed:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, priority, completed, createdAt, updatedAt]
    CreateTodoBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
        dueDate:
          type: string
        completed:
          type: boolean
      required: [title]
    UpdateTodoBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        completed:
          type: boolean
...